20231108015007: Arreglando velocidad,funcionando

+++ src/inc/inc_functions.h
@@ -43,7 +43,7 @@ void getAndSetParamsOne(String InString,     String& option,
 void proccesActionOne(String InString, VectorClassString& vecStrOne);
  
 String getBluetoot(SoftwareSerial &BTSerial,char &charBT);
-unsigned long calcLoopTime(int velocity,int globalVelocity);
+long calcLoopTime(long velocity,long globalVelocity);
 
 
 //int **createMatrix(int rows,int cols);
